Update method signatures and iteration logic in Localization Generators
Changes were made to the way regions are iterated over in the CountriesEnumerableGenerator, changing the type from an IEnumerable to a RegionInfo array. Additionally, methods `GenerateMethod` to `GetExtensionMethodSyntax` in ExtensionsGenerator have been renamed and updated for better clarity and understanding. The trailing trivia was also changed to include line feeds for a more readable code.