Remove unused files in Frank.Analyzers.Tests
This commit removes unused files such as _mingw_dxhelper.h and cgltf.h present in the external glfw and deps folder respectively of the Frank.Analyzers.Tests project. These files contained unnecessary code and definitions which were not being used anywhere in the project. It helps to maintain clean and efficient codebase.